STATE_SAVING_METHOD
- potential remedy for 'strange errors when navigating to login.xhtml`

The Definitive Guide to JSF (2018) Chapter 9 (p. 328):

ViewExpiredException
--------------------
.. [is] thrown when the JSF view state associated with the postback cannot be found
in the HTTP session any more — in other words, when the javax.faces.STATE_SAVING_METHOD
context parameter is set to its default value of “server”, and the end user submits
a JSF form whose view state cannot be found any more on the server side.

Note that when the context parameter is set to “client”, you still need
a “jsf/ClientSideSecretKey” environment entry [TODO, commented out yet] in order to
avoid expired views when the server restarts. This all is elaborated in the section
“View State” in Chapter 3. Also, in the section “@ViewScoped” in Chapter 8, you can find
how to configure the amount of views and managed beans JSF will save in the HTTP session.

There are several circumstances where a ViewExpiredException may unexpectedly
occur. All of them are related to page navigation and the browser cache while the end
user has recently logged out from the web application.